changed IObservable to an abstract class - kept in the interfaces file as it still functions as an interface. this is because the methods relating to IObservable should be the same for the inhereting classes, so an abstract makes more sense than an interface. also, this allows fields to be inherited as well. i still need to write the notifyobservers method (should be easy).
created all the constructors - including overloaded constructor for move
(remains to be seen the usefulness of this). later can add overloaded
constructors for board and movehistory to allow positions to be loaded
in after they have happened (for puzzles, etc). maybe at some point the
Board class can have methods setPosition(MoveHistory): void, and
setStartPosition(): void, which calls the former, to add the pieces into
the list of pieces. for now this functionality is in the board
constructor.

added the squares as a field of the board, which was in the uml before
but overlooked. i have realised that internally, ranks and files will be
0 indexed whereas in real life they are 1 indexed. this is not a huge
issue as the function which converts to algebraic notation can take care
of this. i believe this option is better than 1 indexing internally, as
that would be a recipe for disaster. added the
Square.getAlgebraicPosition() method as i needed it to finish the Board
constructor. this constructor is complaining that it could possibly
through an ArgumentNullException, which is technically true but should
not happen during normal execution. hence, null forgiving operators have
been used here - hopefully this should be okay as i do want the program
to crash with a NullReferenceException if there is an error here.

created ExampleEvaluationStrategy - slightly worried about the fact that
i needed to create an instance of this inside the engine class? although
i believe that is how it is supposed to work, it feels strange

ridiculous piece generation code could be cause for a piecefactory class
- we shall see.