changed the order of some fields and intialisers in gamecontroller. updated the gamecontroller constructor to add the views as observers for the relevant parts of the model.
completed the notifyObservers() method in the Observable class (was very
simple)

baffled by the fact that splashkit does not allow the screen size to be
set dynamically - although slightly grateful as this would be a
nightmare for the screen sizing. created a new class for the global
constant screen size - hopefully can use this to dynamically calculate
the sizes of other components which means changing the screen size
should be an easy task.

broke the interfaces, observable class, and global constants into their
own namespace and gave them a folder in the file system to make it
easier to manage.

added a DrawViews() method to the GameController, to run in the
splashkit game loop. reworked the way that GameController works
(slightly) to have a List<IView> instead of specific fields for each of
the two views (this is more maintainable and extensible in the future)

added a PrintFields method to GlobalSizes for debugging in the future.
created a CusomColors global constants class so that theme changes can
be easily made in the future. i’m worried i have created duplicate
funcionality here with my player colours enum, but I will need advice
from ed regarding that.

i believe, after much consideration and consultation with the internet,
that it would be best to abandon the observer pattern in this case. this
is due to the fact that the game loop updates every frame anyway. this
is a tricky decision and i will certainly need to consult with ed about
whether it is a good idea. on one hand, it doesn’t really seem to make
sense to hold onto this pattern at this point. instead, i should like
the views to hold a reference to their observable counterparts, not the
other way around. this change has been made but i fear that losing
another design pattern has left us with few to talk about. more thinking
is certainly required.

realised all method names need to be in PascalCase, will fix this in the
next commit. i will also be switching to using properties instead of
fields, as i cannot find anything prohibiting this in the unit style
guide and i find the use of fields infuriating when there is a better
option. hopefully the refactoring will be worth it.

unfortunately splashkit does not support the rendering of unicode
symbols such as the chess piece icons. i will need to use images for
this. for now, we have the boardview working using just the names of the
pieces written out instead.