i want to add the ability to click and drag the pieces across the board - later i will add the functionality that means that only legal moves will be allowed, and if possible i wil also add highlighting, time permitting of course.
my plan for this is to add a new property to piece, IsPickedUp, which
will mean that it will always be drawn at the mouse pointer if it is set
to true. otherwise, it will be drawn at the square stored in Location,
as it is already. IsPickedUp will be toggled when clicking the mouse
over a square containing a piece.

i had to add "HandleClick()" to gamecontroller and iview, but now you
can pick up pieces and they will follow the mouse cursor. in reality,
i'm going to make this the behaviour on mousedown, and there will be
separate functions for click and mouseup. to be added in the next
comimt.