refactored using Rider's built in refactoring - made some rearrangements to files.
added new PiecePickedUp property for the controller to track (state pattern?). there are certainly many inefficiencies in this code but i am aiming for fast development time rather than perfect performance (modern systems are more than fast enough to make these computations anyway)

HandleMouseDown for views now takes a gamecontroller object as well - this is so that the function in boardview can call upon the controller’s state to check the PiecePickedUp state

now only one piece can be picked up at a time yay!

—

writing the handleMove() function

changed Move.To to readonly property so it could be used to pattern match moves to legal moves.

removed overloaded constructors for move in favour of using more explicit code to create moves.

—

writing the legal move generator the simplest piece i can think to start with - rooks

created a new orthogonaldirections enum but then deleted it in favour of the implemented form.

removed the promotion field from the Move type, because it can be inferred from if a pawn moves to the end

it appears that the rook movement is working!! and it calculates with negeligable latency which is good because i was ever so slightly worried about the number of foreach loops. i wonder if i should write a function that does Pieces.Find(p=>p.Location=this) in square so i can more easily get the pieces at a square as well as the other way around. perhaps.