made gamecontroller a singleton by adding new Instance readonly property and making the constructor private. this also involved removing the controller param from iview’s handle mouse event functions
created a new IPieceLocation (empty) interface, for Square and the new class NotOnBoard to inherit from. the controller has one NotOnBoard which is where piees are sent after they are taken. the type of Piece.Location is now IPieceLocation instead of Square.

changed my mind about going with this method. instead created a new method on the controller which handles the deletion of pieces from the piece list. this is causing an exception to be thrown by the board view because it is doing some weird stuff with copying the board, i think? i’m gonna change that now that i have the singleton controller class, and see if it helps. this involved changing GameController.Board to a readonly property, and removing the board property from the boardview. drawboard and drawpieces also don’t have any arguments now. unfortunately this did not fix the issue but it has improved the code some - taking a detour to apply this change to the Piece.GetLegalMoves method as well

I foundt the source of the error - yay. I had made a typo when rewriting Square.GetAlgebraicPosition() to use an array instead of a switch - something so inconsequential i didn’t even bother recording it. well, lesson learned.

now i can move on to the creation of the rest of the GetLegalMoves functions.