added hasmoved property to Piece. implemented castling.
added a new readonly prop to kings to check if they are in check.

added Square.IsAttackedBy(PlayerColors player) to calculate checks. also added a readonly prop King.IsInCheck which uses this function to find out if the king is in check

took the initial position setup out of the board constructor so that we can itialise empty boards - this is required for cloning boards for move tests

added a public clone function to the board to create a deep copy of it. this also meant creating a clone function for the piece class, which takes a Board newBoard.

handleMove now clones the board and checks for check before allowing a move to be executed. because of this, squares have a board readonly prop (circular reference bad?). moves also have a clone function which takes in a cloned board (required to create the necessary references).

took some functionality out of handlemove and moved it into Piece.GetLegalMoves(Board board). this method is common to all pieces and removes moves which result in checks from the legal moves after they have been generated by the individual pieces. the method that does this has been renamed to abstract GetPseudoLegalMoves(Board board).

created a piece factory and moved all piece creation in there.

throughout this commit i’ve realised that i made a huge mistake. i thought that global state was a good idea - a rookie error, naturally. i thought using my singleton GameController.Instance everywhere to get board and pieces, etc, would make the code cleaner - not having to pass anything into methods sounded great. how wrong i was - especially with the cloning this introduced so many nasty bugs which were so hard to find. oh well, you live and learn. i’ve removed most of the offending code now, thank good. i’ve allowed some to stay just for views, etc. hopefully i don’t regret it… (ps, this means i’ve removed GameController.HandleCapture).

there’s now Piece.GetAttackedSquares(Board board). The reason for this is that the king casling checks were calling IsInCheck, which was calling Square.IsAttackedBy(), which was calling GetPseudoLegalMoves for every piece, including the king, causing an infinite recusrive call and a stack overflow everytime a piece was moved. By checking the attacked squares instead of the pseudolegal moves, we avoid this error while getting the same result.

everything is working perfectly yay! the pawn code needs some major cleaning up - but i’ll do it when i add en passant. that’ll be after move history is implemented, because we need to check what the precvious move was to know if en passant is allowed.

haven't updated uml yet - coming soon.