adding en passant!
added public Move MostRecentMove property to movehistory for peeking the most recent move. this is required for en passant.

to this end, also had to make the gamecontroller’s movehistory a public readonly property instead of a private field. this is introducing global state again which i might regret later - this time though i have it in mind and for now it’s the easiest way.

doing this i realised that Board was still a public property but could now become a private field. yay

completely restructred the very messy pawn code. it’s lovely now!

added Move.IsDoublePush - required for calculations. also made Move.PieceMoved a readonly prop.

i think it works!