added more theme and global size values.
made MoveHistory.Moves a public readonly prop

i’ve been struggling with an issue for quite some time - isCheck and isCheckmate in Move need to clone the move to execute it on a cloned board in order to determine their truthyness. however, moves can’t be cloned after they have been executed - in general the cloning code is held together by a thread and cloning can only occur at very specific times. the main reason for this is because of the way the cloned moves need to identify which pieces they belong to, but if these pieces have been moved it becomes very difficult. i think the solution to this is that Move.IsCheck, IsCheckmate, etc should be called once in the move constructor, and not be calculated proerties at all. this actually makes more sense, really.

that didn’t end up working - it’s because the constructor was calling clone, which creates a new move, which caused an infinite loop. the best thing to do, i think, is to go back to the lazy method but require the values to be calculated and cached before move execution. this shouldn’t be too hard to implement.

that also didn’t work because unfortunately there was still a recursive call. my new plan to fix this is implement a way to skip the ischeck and ischeckmate calculations entirely for cloned moves. i’ll do this by making a new CloneAndExecute function for Move, and changing the existing Clone to private. then, clone and execute will execute the move surpressing the check for IsCheck and PreventsMovement. To do this properly I also overloaded Execute() with a supressCheckStautusCalculation (default false) to use for cloned moves only.

this is finally working yay